Refactor App.js to update route for Path component
